Model names in extra module; TypeScript compilation created undefined exports symbol (bug?)
