+++ src/app/db/names.ts
@@ -1,0 +1,10 @@
+// Model names enum, because magic strings are bad!
+
+export enum ModelNames {
+	Article = 'Article',
+	EndUser = 'EndUser',
+	Feedback = 'Feedback',
+	Suggestion = 'Suggestion',
+	User = 'User',
+	Website = 'Website',
+}
